refactor(tests): consolidate clients module tests - remove Mock redundancy
Changes:
- Merge test_github_client_prs_review.py into test_github_client_prs.py
  - Moved 6 PR diff/files error handling tests
  - Kept comprehensive auth tests from prs.py (4 tests vs 1 in review)
  - Removed duplicate ReviewMixin fixture
- Delete test_github_client_prs_review.py (171 lines)
- Delete test_issue_close.py (94 lines) - covered by service layer tests
- Remove 2 DI Mock tests from test_github_client_prs.py:
  - test_di_injection_with_mock_client (pure mock chain, no project logic)
  - test_di_injection_in_service_context (mock simulates service, no real logic)

Net result: -206 lines, 19→17 test files in clients module
All 154 clients tests pass, no type/lint errors introduced

Issue: #1006
